clean up latentstyles to use modern attribute implementation
