--- ts/styles/latent-styles/exceptions/index.ts
@@ -1,10 +1,0 @@
-import { XmlComponent } from "../../../docx/xml-components";
-import { LatentStyleExceptionAttributes } from "./attributes";
-
-export class LatentStyleException extends XmlComponent {
-
-    constructor(attributes: LatentStyleExceptionAttributes) {
-        super("w:lsdException");
-        this.root.push(attributes);
-    }
-}
